+++ opal/legacy/api/views/app_general.py
@@ -1,4 +1,5 @@
 """Collection of api views used to display the Opal's General view."""
+from drf_spectacular.utils import OpenApiParameter, extend_schema
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -10,6 +11,19 @@ from opal.patients.models import Relationship, RelationshipStatus
 from ..serializers import AnnouncementUnreadCountSerializer
 
 
+@extend_schema(
+    parameters=[
+        OpenApiParameter(
+            name='Appuserid',
+            location=OpenApiParameter.HEADER,
+            required=True,
+            description='The username of the logged in user',
+        ),
+    ],
+    responses={
+        200: AnnouncementUnreadCountSerializer,
+    },
+)
 class AppGeneralView(APIView):
     """Class to return general page required data."""